feat: 增加CFS缓存大小与数量统计 #1626
