+++ src/backend/common/common-storage/storage-metrics/src/main/kotlin/com/tencent/bkrepo/common/storage/metrics/StorageCacheMetrics.kt
@@ -1,0 +1,70 @@
+/*
+ * Tencent is pleased to support the open source community by making BK-CI 蓝鲸持续集成平台 available.
+ *
+ * Copyright (C) 2024 THL A29 Limited, a Tencent company.  All rights reserved.
+ *
+ * BK-CI 蓝鲸持续集成平台 is licensed under the MIT license.
+ *
+ * A copy of the MIT License is included in this file.
+ *
+ *
+ * Terms of the MIT License:
+ * ---------------------------------------------------
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated
+ * documentation files (the "Software"), to deal in the Software without restriction, including without limitation the
+ * rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to
+ * permit persons to whom the Software is furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all copies or substantial portions of
+ * the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT
+ * LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN
+ * NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
+ * WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
+ * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+
+package com.tencent.bkrepo.common.storage.metrics
+
+import io.micrometer.core.instrument.Gauge
+import io.micrometer.core.instrument.MeterRegistry
+import java.util.concurrent.ConcurrentHashMap
+import java.util.concurrent.atomic.AtomicLong
+
+class StorageCacheMetrics(
+    private val registry: MeterRegistry,
+) {
+
+    private val cacheSizeMap = ConcurrentHashMap<String, AtomicLong>()
+    private val cacheCountMap = ConcurrentHashMap<String, AtomicLong>()
+
+    /**
+     * 设置当前缓存总大小，由于目前只有Job服务在清理缓存时会统计，因此只有Job服务会调用该方法
+     */
+    fun setCacheSize(storageKey: String, size: Long) {
+        val s = cacheSizeMap.getOrPut(storageKey) { AtomicLong(size) }
+        s.set(size)
+        Gauge.builder(CACHE_SIZE, s) { it.toDouble() }
+            .tag("storageKey", storageKey)
+            .description("storage cache total size")
+            .register(registry)
+    }
+
+    /**
+     * 设置当前缓存总数，由于目前只有Job服务在清理缓存时会统计，因此只有Job服务会调用该方法
+     */
+    fun setCacheCount(storageKey: String, count: Long) {
+        val c = cacheCountMap.getOrPut(storageKey) { AtomicLong(count) }
+        c.set(count)
+        Gauge.builder(CACHE_COUNT, c) { it.toDouble() }
+            .tag("storageKey", storageKey)
+            .description("storage cache total count")
+            .register(registry)
+    }
+
+    companion object {
+        private const val CACHE_SIZE = "storage.cache.size"
+        private const val CACHE_COUNT = "storage.cache.count"
+    }
+}
