+++ src/backend/common/common-storage/storage-metrics/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,2 @@
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+com.tencent.bkrepo.common.storage.StorageAutoConfiguration
